feat: Added event handlers and relocated events located in the main index.ts into their own respective files
