+++ src/discord/events/index.ts
@@ -1,0 +1,5 @@
+import ReadyEventHandler from './readyEventHandler';
+import CommandInteractionCreateEventHandler from './commandInteractionCreateEventHandler';
+import MessageCreateEventHandler from './messageCreateEventHandler';
+
+export default [ReadyEventHandler, CommandInteractionCreateEventHandler, MessageCreateEventHandler];
